hw3/backend/requirements.txt
black==24.8.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
psycopg2-binary